+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.2
+- Added an option in preferences to confirm closing the tab so users can avoid accidentally losing all their progress.
+- Updated the `Language` section of the preferences sidebar to more accurately reflect the user's language setting.
+- Rewrote `modules/config` to prevent dynamic indexing while also auto-saving via a custom setter.
+- Version display now links to the changelog.
+
 # 1.0.1
 - Removed debug window variables.
 - Removed `landmark` from the icon menu.